Renaming effect_registry.py to effects.py
