+++ pyminion/effects.py
